requirements.txt
flet
supabase
python-dotenv